perf(phaser): simplify participant state and reuse task wakers
