+++ benchmarks/Cargo.toml
@@ -39,6 +39,7 @@ asyncband = { workspace = true, features = [
   "once-cell",
   "once-map",
   "oneshot",
+  "phaser",
   "pool",
   "rwlock",
   "semaphore",
