+++ xtask/src/main.rs
@@ -121,6 +121,10 @@ impl CommandMiri {
             &["--test", "unsafe_paths_test"],
         ));
         run_command(make_miri_cmd("tests-integration", &["--test", "mpsc_test"]));
+        run_command(make_miri_cmd(
+            "tests-integration",
+            &["--test", "phaser_test"],
+        ));
     }
 }
